Splitting dashboard provisioning in two disk and vcs
