+++ pkg/services/provisioning/dashboards/disk/config_reader.go
@@ -1,4 +1,4 @@
-package dashboards
+package disk
 
 import (
 	"fmt"
