+++ pkg/services/provisioning/dashboards/disk/config_reader_test.go
@@ -1,4 +1,4 @@
-package dashboards
+package disk
 
 import (
 	"errors"
