+++ pkg/services/provisioning/dashboards/disk/dashboard_mock.go
@@ -1,4 +1,4 @@
-package dashboards
+package disk
 
 import "context"
