+++ pkg/services/provisioning/dashboards/disk/file_reader_linux_test.go
@@ -1,6 +1,6 @@
 // +build linux
 
-package dashboards
+package disk
 
 import (
 	"path/filepath"
