+++ pkg/services/provisioning/dashboards/disk/testdata/test-configs/dashboards-from-disk/sample.yaml
