+++ pkg/services/provisioning/dashboards/disk/testdata/test-dashboards/folders-from-files-structure/folderTwo/dashboard2.json
