+++ pkg/services/provisioning/dashboards/disk/testdata/test-dashboards/one-dashboard/dashboard1.json
