+++ pkg/services/provisioning/dashboards/disk/testdata/test-dashboards/symlink
