+++ pkg/services/provisioning/dashboards/disk/types.go
@@ -1,4 +1,4 @@
-package dashboards
+package disk
 
 import (
 	"fmt"
